ffi: pass osc arguments as list of void * to callback
The list of arguments is now converted to void * before being passed to
the callback. This is done, because C can't understand the complex type
that is `OscType`, because it is a Rust enum. So, before this commit
the C callback could not have possibly used the type argument.

To fix this, everything is now converted to void * and a new argument
`types` is introduced, which is a string of all types that get passed.

It still needs to be tested, how slow preparing the arguments for C is.
I expect it to be fine because it is only a few allocations and network
is the slowest thing by far in the chain.